[shared] cleanup serdes.ipc
