+++ integration_tests/test_suites/daemon-test-suite/utils.py
@@ -1,6 +1,6 @@
 import contextlib
 
-from dagster_shared.serdes.ipc import interrupt_ipc_subprocess, open_ipc_subprocess
+from dagster_shared.ipc import interrupt_ipc_subprocess, open_ipc_subprocess
 
 
 @contextlib.contextmanager
