+++ python_modules/dagster/dagster/_daemon/cli/__init__.py
@@ -4,7 +4,7 @@ from contextlib import ExitStack
 from typing import Optional
 
 import click
-from dagster_shared.serdes.ipc import interrupt_on_ipc_shutdown_message
+from dagster_shared.ipc import interrupt_on_ipc_shutdown_message
 
 from dagster import __version__ as dagster_version
 from dagster._cli.utils import assert_no_remaining_opts, get_instance_for_cli
